% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ab_ahus}
\alias{ab_ahus}
\title{AHUS antibiotics list}
\format{
\describe{
\item{Antibiotic}{Name of the antibiotic}
\item{Category}{Access, watch or reserve}
\item{Norsk_AHUS}{The Norwegian name, in AHUS hospital system}
\item{color_code}{Color code according to category}
}
}
\usage{
ab_ahus
}
\description{
This dataset contains information on antibiotics from AHUS. More details to follow.
}
\examples{
head(ab_ahus)
}
\keyword{datasets}
