% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ab_prescription}
\alias{ab_prescription}
\title{Demo data: antibiotics prescription}
\format{
\describe{
\item{ID}{Unique identifier}
\item{time}{Record time (in hours)}
\item{ab_name}{Name of the antibiotics}
\item{firsttime}{Whether it is the first time this antibiotics is prescribed}
\item{abform_filled}{Whether the form for antibiotics is filled}
\item{purpose}{Purpose for the antibiotics prescription}
\item{source}{Type of infection: community acquire, hospital acquired or profylactic uses}
}
}
\usage{
ab_prescription
}
\description{
This dataset contains 10 samples of randomly generated data to for demonstration purpose.
}
\examples{
head(ab_prescription)
}
\keyword{datasets}
