% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ab_use}
\alias{ab_use}
\title{Demo data: antibiotics use}
\format{
\describe{
\item{ID}{Unique identifier}
\item{time}{Record time (in hours)}
\item{use}{Dose of antibiotics}
\item{unit}{Unit of the drug}
\item{ab_name}{Name of the antibiotics}
\item{ab_method}{Method of drug use}
\item{ab_dose}{Additional information of the drug use}
}
}
\usage{
ab_use
}
\description{
This dataset contains 10 samples of randomly generated data to for demonstration purpose.
}
\examples{
head(ab_use)
}
\keyword{datasets}
