% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ab_who}
\alias{ab_who}
\title{WHO antibiotics list}
\format{
\describe{
\item{Antibiotic}{Name of the antibiotic}
\item{Class}{Antibiotic class}
\item{ATC code}{ATC code}
\item{Category}{Access, watch or reserve}
\item{Listed on EML/EMLc 2023}{Whether listed}
\item{color_code}{Color code according to category}
}
}
\usage{
ab_who
}
\description{
This dataset contains information on antibiotics from WHO. More details to follow.
}
\examples{
head(ab_who)
}
\keyword{datasets}
