% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{breakpoints}
\alias{breakpoints}
\title{Breakpoint}
\usage{
breakpoints(x, n = 8)
}
\arguments{
\item{x}{time vector}

\item{n}{number of breakpoints}
}
\value{
a vector of breakpoints
}
\description{
Make pretty break points for plotting
}
