% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{catheters}
\alias{catheters}
\title{Demo data: catheters}
\format{
\describe{
\item{ID}{Unique identifier}
\item{time}{Record time (in hours)}
\item{catheter}{Name of the catheter}
\item{use}{How long the catheter has been used}
\item{unit}{Minutes}
}
}
\usage{
catheters
}
\description{
This dataset contains 10 samples of randomly generated data to for demonstration purpose.
}
\examples{
head(catheters)
}
\keyword{datasets}
