% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Demo data: demographics}
\format{
\describe{
\item{ID}{Unique identifier}
\item{sex}{Sex of the patient}
\item{age}{Age}
\item{dept}{Department of admission: gastrosurgery, orthopaedics, gastro medicine, neurology}
\item{t0}{Admission time (in hours). This is relative to the data extraction time.}
\item{los}{Length of stay (in hours)}
\item{adm_type}{Type of admission: emergency or elective}
}
}
\usage{
demographics
}
\description{
This dataset contains 10 samples of randomly generated data to for demonstration purpose.
}
\examples{
head(demographics)
}
\keyword{datasets}
