% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{f1}
\alias{f1}
\title{A simple placehold function}
\usage{
f1(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a value 3 greater than the input
}
\description{
A simple placehold function
}
\examples{
f1(5)
}
