% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{loc_expand}
\alias{loc_expand}
\title{Expand location}
\usage{
loc_expand(x)
}
\arguments{
\item{x}{data}
}
\value{
A dataframe with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
