% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{location}
\alias{location}
\title{Demo data: location}
\format{
\describe{
\item{ID}{Unique identifier}
\item{postid}{ID for the stay of the major ward}
\item{subpostid}{ID for the stay of the associated ward: only sop (central operation room), intensive (intensive care unit) and postop (post operative unit)}
\item{post}{Name of the ward}
\item{time}{time}
\item{time_stay_minutes}{Minutes}
}
}
\usage{
location
}
\description{
This dataset contains 10 samples of randomly generated data to for demonstration purpose.
}
\examples{
head(location)
}
\keyword{datasets}
