% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{location_meta}
\alias{location_meta}
\title{Location meta-data}
\format{
\describe{
\item{location_code}{Location code}
\item{location_display}{Location name to display. These are abbreviations}
\item{type}{Type of location: ER, Postoperative, ICU or Bedward}
}
}
\usage{
location_meta
}
\description{
This dataset contains information on locations at AHUS hospital. This is only
}
\examples{
head(location_meta)
}
\keyword{datasets}
