% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{make_demographic_info}
\alias{make_demographic_info}
\title{Make demographic info}
\usage{
make_demographic_info(
  demo_df,
  keep_id = T,
  keep_sex = T,
  keep_age = T,
  keep_dept = T,
  keep_admtype = T,
  keep_t0 = T,
  keep_los = T
)
}
\arguments{
\item{demo_df}{drug dataframe}

\item{keep_id}{tmin}

\item{keep_sex}{tmax}

\item{keep_age}{description}

\item{keep_dept}{t}

\item{keep_admtype}{t}

\item{keep_t0}{t}

\item{keep_los}{t}
}
\value{
A datafrme with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
