% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{make_event_drug_use}
\alias{make_event_drug_use}
\title{Make drug use events}
\usage{
make_event_drug_use(drug_df, tmin = NULL, tmax = NULL)
}
\arguments{
\item{drug_df}{drug dataframe}

\item{tmin}{tmin}

\item{tmax}{tmax}
}
\value{
A datafrme with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
