% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{make_location}
\alias{make_location}
\title{Make location}
\usage{
make_location(data_location, tadmin, los)
}
\arguments{
\item{data_location}{drug dataframe}

\item{tadmin}{tmin}

\item{los}{tmax}
}
\value{
A datafrme with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
