% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_drug_prescribe}
\alias{plot_drug_prescribe}
\title{Plot drug prescription}
\usage{
plot_drug_prescribe(data_prescribe, plot_obj)
}
\arguments{
\item{data_prescribe}{drug prescription dataframe}

\item{plot_obj}{plot object from drug use}
}
\value{
An enhanced plot with drug prescription events
}
\description{
Plot drug prescription events over time, on top of drug use plot. It is necessary to supply a drug use plot first.
}
