% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_drug_prescribe_interactive}
\alias{plot_drug_prescribe_interactive}
\title{Make drug use events}
\usage{
plot_drug_prescribe_interactive(data_prescribe, plot_obj)
}
\arguments{
\item{data_prescribe}{drug dataframe}

\item{plot_obj}{tmin}
}
\value{
A datafrme with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
