% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_drug_use}
\alias{plot_drug_use}
\title{Plot drug use over time}
\usage{
plot_drug_use(data_drug_use)
}
\arguments{
\item{data_drug_use}{drug dataframe}
}
\value{
A plot that visualizes drug use events over time
}
\description{
Plot drug use over time. For antibiotics, a tricolor systme (WHO AWaRe) is applied.
}
