% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_info_demographic}
\alias{plot_info_demographic}
\title{Plot demographic information card}
\usage{
plot_info_demographic(info_text, plot_obj)
}
\arguments{
\item{info_text}{Text information, parsed to be plotted. See make_demographic_info() for more.}

\item{plot_obj}{plot object from drug use}
}
\value{
An enhanced plot that has demographic information
}
\description{
Plot demographic information card
}
