% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_location}
\alias{plot_location}
\title{Make drug use events}
\usage{
plot_location(loc_obj, keep_text = T, keep_legend = F, keep_time = T)
}
\arguments{
\item{loc_obj}{drug dataframe}

\item{keep_text}{tmin}

\item{keep_legend}{tmax}

\item{keep_time}{description}
}
\value{
A dataframe with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
