% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{x}
\alias{x}
\title{Placeholder data x}
\format{
\describe{
\item{x}{The placeholder data x}
}
}
\usage{
x
}
\description{
This dataset contains one value, x
}
\examples{
print(x)
}
\keyword{datasets}
