% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{breakpoints}
\alias{breakpoints}
\title{Make drug use events}
\usage{
breakpoints(x, n = 8)
}
\arguments{
\item{x}{drug dataframe}

\item{n}{tmin}
}
\value{
A datafrme with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
