% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{loc_expand}
\alias{loc_expand}
\title{Expand location}
\usage{
loc_expand(x)
}
\arguments{
\item{x}{location dataframe}
}
\value{
a datatable with timestamps and location code
}
\description{
Expand location for the plot as color-blocks. Typically used inside function: make_location()
}
