% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{make_demographic_info}
\alias{make_demographic_info}
\title{Make demographic info}
\usage{
make_demographic_info(
  demo_df,
  keep_id = T,
  keep_sex = T,
  keep_age = T,
  keep_dept = T,
  keep_admtype = T,
  keep_t0 = T,
  keep_los = T
)
}
\arguments{
\item{demo_df}{Demographics dataframe}

\item{keep_id}{Whether to keep ID. Logical, default is True}

\item{keep_sex}{Whether to keep sex. Logical, default is True}

\item{keep_age}{Whether to keep age. Logical, default is True}

\item{keep_dept}{Whether to keep department. Logical, default is True}

\item{keep_admtype}{Whether to keep admission type. Logical, default is True}

\item{keep_t0}{Whether to keep T0. Logical, default is True}

\item{keep_los}{Whether to keep LOS (length of stay). Logical, default is True}
}
\value{
A parsed string with demographic information
}
\description{
Prepare data: demographics for the static information card
}
