% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_functions.R
\name{make_location}
\alias{make_location}
\title{Make location}
\usage{
make_location(data_location, tadmin, los)
}
\arguments{
\item{data_location}{location data}

\item{tadmin}{Time of admission}

\item{los}{Length of stay (in hours)}
}
\value{
A dataframe with location
}
\description{
Prepare data for the plot: location
}
