% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_drug_prescribe_interactive}
\alias{plot_drug_prescribe_interactive}
\title{Plot drug prescription interactively with additional information}
\usage{
plot_drug_prescribe_interactive(data_prescribe, plot_obj)
}
\arguments{
\item{data_prescribe}{drug prescription dataframe}

\item{plot_obj}{plot object from drug use}
}
\value{
An enhanced plot with drug prescription events
}
\description{
Plot drug prescription interactively with additional information
}
