% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_drug_use}
\alias{plot_drug_use}
\title{Make drug use events}
\usage{
plot_drug_use(data_drug_use)
}
\arguments{
\item{data_drug_use}{drug dataframe}
}
\value{
A plot
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
