% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_info_demographic}
\alias{plot_info_demographic}
\title{Make drug use events}
\usage{
plot_info_demographic(info_text, plot_obj)
}
\arguments{
\item{info_text}{drug dataframe}

\item{plot_obj}{tmin}
}
\value{
A datafrme with drug uses
}
\description{
Collect results from the new diet discovery, along with the current diet.
}
