% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_location}
\alias{plot_location}
\title{Plot location}
\usage{
plot_location(loc_obj, keep_text = T, keep_legend = F, keep_time = T)
}
\arguments{
\item{loc_obj}{location object}

\item{keep_text}{Keep text. Logical, default is True}

\item{keep_legend}{Keep legend. Logical, default is True}

\item{keep_time}{Keep time. Logical, default is True}
}
\value{
A plot of location information
}
\description{
Plot location information as color blocks
}
